Make call for action optional (#50)
* feat: Make call for actions optional

Allow configuration to enable or disable the call for action (CTA) section on
the front page.

BREAKING CHANGE: `enable` must be defined to `true` in the `content/_index.md`
file to enable CTA section on the front page. It's no longer enabled by
default.

Signed-off-by: Julien Riou <julien@riou.xyz>

* style: Generate CTA heading & message if needed

Call for action may not have heading nor message. Do not create title or
paragraph if they are not configured. This avoids useless empty space.

Signed-off-by: Julien Riou <julien@riou.xyz>

---------

Signed-off-by: Julien Riou <julien@riou.xyz>