Adds map and map_marker shortcodes (#58)
The map on the homepage uses the same render code, but still currently
gets added by default.

This does introduce a breaking change. The code to include a map on a
page has changed from the old

```
{{< loc.inline >}}{{ partial "map" . }}{{< /loc.inline >}}
```

Using the `map` and `map_marker` shortcodes gives more flexibility, but
there is a change required.

ALso, I have only been able to verify it works using the OSM map
provider.